SQL statements for loading into a metadata database.
We have data in current metadata dbs, which we want to import
into a database with the new schema. For many tables this is
fairly straightforward, as they have not fundamentally changed.
And for assemblies and genesets, there is enough information
in the current metadata db to create the corresponding genome
and dataset objects. However, for other datasets, the metadata
db is insufficient, and we need to retrieve data from the core
(and corelike) databases. In addition, we want a different set
of geneset attributes in the new metadata db, which also need
to be extracted from the core dbs.

The SQL to extract data from the core dbs will be the basis
of the SQLalchemy-based code that will run when a database is
handed over. For now, the 'datasets_*.sql' and 'attributes_*.sql'
files need to be run against the server with the core databases,
with the result files loaded into the two tables defined in
'tmp_tables.sql'. Then 'load_metadata.sql' (currently hard-coded
to load from 'ensembl_metadata_qrp') will populate a metadata
db of the new schema.