Load assembly_sequences from core dbs.
The problem with loading assembly_sequences from an existing
metadata db is that the information is incomplete or inconsistent.
We need to be a bit smart about assigning the accession and name
to the correct field, determine the sequence location and
chromosomal status, and calculate a checksum. We get all the data
from the cores into a single file, which is then imported into
the metadata db.